Platinum Stack bj3015_오아시스재결합

+++ Java/Baekjoon/Platinum/bj3015_오아시스재결합.java
@@ -9,20 +9,20 @@ public class bj3015_오아시스재결합 {
     public static void main(String[] args) throws Exception {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int N = Integer.parseInt(br.readLine());
-        Deque<long[]> stack = new ArrayDeque<>();
+        Deque<int[]> stack = new ArrayDeque<>();
         long answer = 0;
         for (int i = 0; i < N; i++) {
             int n = Integer.parseInt(br.readLine());
-            long cnt = 0;
-            long same = 1;
+            int cnt = 0;
+            int same = 1;
             while (!stack.isEmpty() && stack.peekFirst()[0] <= n) {
-                long[] pop = stack.pop();
+                int[] pop = stack.pop();
                 cnt += pop[1];
                 if (pop[0] == n) same = pop[1] + 1;
             }
             cnt += !stack.isEmpty() ? 1 : 0;
             answer += cnt;
-            stack.push(new long[]{n, same});
+            stack.push(new int[]{n, same});
         }
         System.out.println(answer);
     }
